Move TestDeletePublicationsClass to dynamic data tests
